Module iced_aw::core::date

Helper functions for calculating dates

Structs

Date

The date value

MAX_MONTH_STR_LEN

Gets the length of the longest month name.

WEEKDAY_LABELS

Gets the labels of the weekdays containing the first two characters of the weekdays.

Enums

IsInMonth

Specifies if the calculated day lays in the previous, same or next month of the date.

Functions

month_as_string

Gets the string representation of the month of the given date.

position_to_day

Calculates the day number at the given position in the calendar table based on the given year and month.

pred_day

Calculates a date with the previous day based on the given date.

pred_month

Creates a date with the previous month based on the given date.

pred_week

Calculates a date with the previous week based on the given date.

pred_year

Creates a date with the previous year based on the given date.

succ_day

Calculates a date with the next day based on the given date.

succ_month

Creates a date with the next month based on given date.

succ_week

Calculates a date with the next week based on the given date.

succ_year

Creates a date with the next year based on the given date.

year_as_string

Gets the string representation of the year of the given date.