Module iced_aw::native

Stateless, pure widgets for iced

Re-exports

pub use badge::Badge;

pub use number_input::NumberInput;

pub use card::Card;

pub use color_picker::ColorPicker;

pub use crate::native::cupertino::cupertino_spinner::CupertinoSpinner;

pub use date_picker::DatePicker;

pub use selection_list::List;

pub use selection_list::SelectionList;

pub use floating_element::FloatingElement;

pub use grid::Grid;

pub use icon_text::IconText;

pub use modal::Modal;

pub use split::Split;

pub use tab_bar::TabBar;

pub use tab_bar::TabLabel;

pub use tabs::TabBarPosition;

pub use tabs::Tabs;

pub use time_picker::TimePicker;

pub use wrap::Wrap;

pub use spinner::Spinner;

Modules

badge

Use a badge for color highlighting important information.

card

Displays a Card.

color_picker

Use a color picker as an input element for picking colors.

cupertino

Cupertino-style widgets

date_picker

Use a date picker as an input element for picking dates.

floating_element

Use a floating element to overlay an element over some content

grid

Use a grid as an input element for creating grids.

icon_text

Text widget for rendering icons.

menu

A MenuBar widget for displaying MenuTrees

modal

A modal for showing elements as an overlay on top of another.

number_input

Display fields that can only be filled with numeric type.

overlay

Display interactive elements on top of other widgets.

quad

A dummy widget that draws a quad

selection_list

Display a dropdown list of selectable values.

spinner

A spinner to suggest something is loading.

split

Use a split to split the available space in two parts to display two different elements.

tab_bar

Displays a TabBar to select the content to be displayed.

tabs

Displays a Tabs widget to select the content to be displayed.

time_picker

Use a time picker as an input element for picking times.

wrap

A widget that displays its children in multiple horizontal or vertical runs.